requirements.txt
# streamlit run app.py

streamlit
pymongo==4.4.0
python-dotenv==1.0.0
google-generativeai==0.2.0